+++ asgardpy/config/__init__.py
@@ -1,3 +1,11 @@
-from asgardpy.config.generator import AsgardpyConfig
+from asgardpy.config.generator import (
+    AsgardpyConfig,
+    get_model_template,
+    recursive_merge_dicts,
+)
 
-__all__ = ["AsgardpyConfig"]
+__all__ = [
+    "AsgardpyConfig",
+    "get_model_template",
+    "recursive_merge_dicts",
+]