+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Combine the various Sky Position configs into a single SkyPositionConfig by [@chaimain](https://github.com/chaimain).
+- Add script to get most preferred spectral model fit by [@chaimain](https://github.com/chaimain).
 
 ## [v0.3.3](https://github.com/chaimain/asgardpy/releases/tag/v0.3.3) - 2023-06-20